pulls out substitute, bumps some versions
