+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "weresocool"
-version = "0.4.0"
+version = "0.5.0"
 authors = ["Danny Meyer <Danny.Meyer@gmail.com>"]
 description = "***** WereSoCool __!Now In Stereo!__ ****** Make cool sounds. Impress your friends."
 license = "MIT"
