+++ parser/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "socool_parser"
-version = "0.2.0"
+version = "0.5.0"
 authors = ["Danny Meyer <Danny.Meyer@gmail.com>"]
 edition = "2018"
 build = "build.rs" # LALRPOP preprocessing
