+++ socool_ast/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "socool_ast"
-version = "0.1.0"
+version = "0.5.0"
 authors = ["Danny Meyer <Danny.Meyer@gmail.com>"]
 edition = "2018"
