+++ songs/wip/lorenzo.socool
@@ -1,0 +1,73 @@
+{ f: 220, l: 1, g: 1, p: 0}
+
+overtones = {
+    O[
+    	(11/4, 8, 1, 1/3),
+    	(11/4, 0, 1, -1/3),
+    	(3, 8, 0, 1/3),
+    	(3, 0, 1, -1/3),
+    	(3/2, 3, 1, 1),
+      	(1/1, 4, 1, 1/2),
+      	(1/2, 0, 1, -1/2),
+      	(1/2, 4, 1, 1/2),
+  	]
+}
+
+thing1 = {
+	Sequence [
+		AsIs, Tm 3/2, Tm 9/8, Tm 6/5
+	]  
+}
+
+thing2 = {
+	Sequence [
+		AsIs, 
+		Tm 7/8, 
+		Tm 4/5,
+	]
+}
+
+thing3 = {
+	Sequence [
+		AsIs, 
+		Tm 4/3, 
+		-- Tm 7/8,
+		-- Tm 6/5
+	]	
+}
+
+unit = { Length 4 }
+
+
+bass = {
+	Overlay [
+		AsIs, 
+		Tm 5 | Gain 1/4 
+	]
+	| Sequence [
+		AsIs, 
+		Tm 3/2, 
+		Tm 5/3, 
+		Tm 3/2, 
+		Tm 7/6
+	] 
+	| Tm 1/2
+	| Sequence [AsIs, Tm 7/8]
+
+}
+
+main = {
+	Overlay [
+		Gain 1/6, 
+		-- overtones | 
+		Sequence [
+			thing1 
+				| thing2 
+				| thing3,
+		] > FitLength unit,
+		bass > FitLength unit
+
+
+	]
+}
+
